requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=8.0.0,<9.0
pytest-django>=4.10.0,<5.0
pytest-cov>=6.0.0,<7.0

# Code quality
ruff>=0.8.0,<1.0
mypy>=1.15.0,<2.0
django-stubs>=5.0.0,<6.0
djangorestframework-stubs>=3.15.0,<4.0

# Monitoring
flower>=2.0.1,<3.0

# Pre-commit
pre-commit>=4.0.0,<5.0
